refactor: change private users endpoint to /signup and /signin
